// wp-content/plugins/manga-overlay-core/editor-src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        ReadingProgressUpdate: {
            chapter_id: number;
            page_index: number;
            progress_unit: number;
            /** @enum {string} */
            reader_mode: "webtoon" | "paged";
        };
// ...
        ReadingProgress: {
            chapter_id: number;
            page_index: number;
            progress_unit: number;
            /** @enum {string} */
            reader_mode: "webtoon" | "paged";
            /** Format: date-time */
            updated_at: string;
        };
// ...
    };
// ...
}